Move installation utilities to compilation module
