+++ pysr/__init__.py
@@ -7,7 +7,7 @@ from .sr import (
     best_callable,
     best_row,
 )
-from .julia_helpers import install
+from .package_compiler import install
 from .feynman_problems import Problem, FeynmanProblem
 from .export_jax import sympy2jax
 from .export_torch import sympy2torch
